se termina de arreglar los reales. Se implementa funcion myAtof, formateoNumero y se pasan los floats y doubles por un tema de precision

+++ y.tab.h
@@ -143,7 +143,7 @@ typedef union YYSTYPE
 
 
 	int valor_int;
-	float valor_float;
+	double valor_float;
 	char *valor_string;
